refactor(usecases) :hammer: refactor of files
